WIP, see next commit

+++ campus/views/views_clubs.py
@@ -1,6 +1,7 @@
 import os
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required
+from django.utils.decorators import method_decorator
 from django.http import HttpResponseRedirect, HttpResponse, Http404
 from django.utils.translation import ugettext_lazy as _
 from django.contrib import messages
@@ -14,7 +15,7 @@ from PIL import Image
 
 from campus.forms import SendMailForm, ClubManagementForm
 from campus.models.clubs_models import StudentOrganisation
-from gestion_personnes.models import LdapUser
+from gestion_personnes.models import LdapUser, LdapGroup
 from fonctions.decorators import ae_required
 
 from myresel.settings import MEDIA_ROOT, LDAP_DN_PEOPLE
@@ -28,6 +29,7 @@ def list_clubs(request):
     assos = [o for o in organisations if "tbAsso" in o.object_classes]
     lists = [o for o in organisations if "tbCampagne" in o.object_classes]
 
+
     return render(
         request,
         'campus/clubs/list.html',
@@ -35,17 +37,26 @@ def list_clubs(request):
             'clubs': clubs,
             'assos': assos,
             'lists': lists,
+
             'ldapOuPeople': LDAP_DN_PEOPLE,
         }
     )
 
-# TODO : gestion des droits
+# TODO : gestion des droits des prezs
 class NewClub(FormView):
+
     template_name = 'campus/clubs/new_club.html'
     form_class = ClubManagementForm
     success_url = '/campus/clubs'
 
+    @method_decorator(login_required)
+    def dispatch(self, *args, **kwargs):
+        return super(NewClub, self).dispatch(*args, **kwargs)
+
     def form_valid(self, form):
+        if not request.ldap_user.is_campus_moderator():
+            messages.error(request, _("Vous n'êtes pas modérateur campus"))
+            return HttpResponseRedirect(reverse('campus:clubs:list'))
         if form.cleaned_data['type'] != "CLUB":
             logo = form.cleaned_data['logo']
             logo = Image.open(BytesIO(logo.read()))
@@ -63,11 +74,19 @@ class NewClub(FormView):
 
 # TODO : gestion des droits
 class EditClub(FormView):
+
     template_name = 'campus/clubs/new_club.html'
     form_class = ClubManagementForm
     success_url = '/thanks/'
 
+    @method_decorator(login_required)
+    def dispatch(self, *args, **kwargs):
+        return super(EditClub, self).dispatch(*args, **kwargs)
+
     def get(self, request, pk):
+        if not request.ldap_user.is_campus_moderator():
+            messages.error(request, _("Vous n'êtes pas modérateur campus"))
+            return HttpResponseRedirect(reverse('campus:clubs:list'))
         try:
             orga = StudentOrganisation.filter(cn=pk)[0]
         except IndexError:
@@ -90,11 +109,18 @@ class EditClub(FormView):
         return render(request, self.template_name, {'form': form, 'pk': pk})
 
     def form_valid(self, form):
+        if not request.ldap_user.is_campus_moderator():
+            messages.error(request, _("Vous n'êtes pas modérateur campus"))
+            return HttpResponseRedirect(reverse('campus:clubs:list'))
         form.edit_club()
         return super(EditClub, self).form_valid(form)
 
 class DeleteClub(View):
+
     def get(self, request, pk):
+        if not request.ldap_user.is_campus_moderator():
+            messages.error(request, _("Vous n'êtes pas modérateur campus"))
+            return HttpResponseRedirect(reverse('campus:clubs:list'))
         try:
             StudentOrganisation.get(cn=pk).delete()
             return redirect('campus:clubs:list')
@@ -112,31 +138,41 @@ class SearchClub(View):
         context={'clubs': clubs}
         return render(request, self.template_name, context)
 
+#TODO gérer l'ajout par un modo campus en methode post (si pas trouvé mieux en attendant)
 class AddPersonToClub(View):
 
-    def get(self, request, pk, user=None):
+    @method_decorator(login_required)
+    def dispatch(self, *args, **kwargs):
+        return super(AddPersonToClub, self).dispatch(*args, **kwargs)
 
+    def get(self, request, pk):
         try:
             club=StudentOrganisation.get(cn=pk)
         except ObjectDoesNotExist:
             raise Http404
-        if user == None:
-            user = request.ldap_user
-        else:
-            if not request.ldap_user.is_campus_moderator():
-                messages.error(request, _("Vous n'êtes pas modérateur campus"))
-                return HttpResponseRedirect(reverse('campus:clubs:list'))
-            else:
-                try:
-                    LdapUser.get(uid=user.uid)
-                except ObjectDoesNotExist:
-                    raise Http404
+        user = request.ldap_user
         if "tbAsso" not in club.object_classes and user.pk not in club.members:
             club.members.append(user.pk)
             club.save()
         return redirect('campus:clubs:list')
 
+    def post(self, request, pk):
+        if not request.ldap_user.is_campus_moderator():
+            messages.error(request, _("Vous n'êtes pas modérateur campus"))
+            return HttpResponseRedirect(reverse('campus:clubs:list'))
+        else:
+            uid = request.POST.get("id_user", "")
+            try:
+                LdapUser.get(uid=uid)
+            except ObjectDoesNotExist:
+                raise Http404
+        #TO BE CONTINUED -->
+
 class RemovePersonFromClub(View):
+
+    @method_decorator(login_required)
+    def dispatch(self, *args, **kwargs):
+        return super(RemovePersonFromClub, self).dispatch(*args, **kwargs)
 
     def get(self, request, pk, user=None):
         try:
@@ -145,16 +181,19 @@ class RemovePersonFromClub(View):
             raise Http404
         if user == None:
             user = request.ldap_user
-        else:
-            if not request.ldap_user.is_campus_moderator():
-                messages.error(request, _("Vous n'êtes pas modérateur campus"))
-                return HttpResponseRedirect(reverse('campus:clubs:list'))
-            else:
-                try:
-                    LdapUser.get(uid=user.uid)
-                except ObjectDoesNotExist:
-                    raise Http404
         if "tbAsso" not in club.object_classes and user.pk in club.members:
             club.members.remove(user.pk)
             club.save()
         return redirect('campus:clubs:list')
+
+        def post(self, request, pk):
+            if not request.ldap_user.is_campus_moderator():
+                messages.error(request, _("Vous n'êtes pas modérateur campus"))
+                return HttpResponseRedirect(reverse('campus:clubs:list'))
+            else:
+                uid = request.POST.get("id_user", "")
+                try:
+                    LdapUser.get(uid=uid)
+                except ObjectDoesNotExist:
+                    raise Http404
+            #TO BE CONTINUED -->
